+++ src/FilePlink.cpp
@@ -77,7 +77,7 @@ void FileBed::read_all_and_centering()
     inbed.shrink_to_fit();
 }
 
-void FileBed::read_snp_block_initial(uint start_idx, uint stop_idx, bool standardize)
+void FileBed::read_snp_block_initial(uint64 start_idx, uint64 stop_idx, bool standardize)
 {
     uint actual_block_size = stop_idx - start_idx + 1;
     // if G is not initial then initial it
@@ -176,7 +176,7 @@ void FileBed::read_snp_block_initial(uint start_idx, uint stop_idx, bool standar
 
 }
 
-void FileBed::read_snp_block_update(uint start_idx, uint stop_idx, const MatrixXf& U, const VectorXf& svals, const MatrixXf& VT, bool standardize)
+void FileBed::read_snp_block_update(uint64 start_idx, uint64 stop_idx, const MatrixXf& U, const VectorXf& svals, const MatrixXf& VT, bool standardize)
 {
     uint actual_block_size = stop_idx - start_idx + 1;
     if (G.cols() < params.blocksize || (actual_block_size < params.blocksize))
